+++ nodes/OKRunit/resources/comment/create.ts
@@ -19,27 +19,5 @@ export const commentCreateDescription: INodeProperties[] = [
 			show: showOnlyForCommentCreate,
 		},
 		description: 'The text content of the comment',
-		routing: {
-			send: {
-				type: 'body',
-				property: 'body',
-			},
-		},
-	},
-	{
-		// Always send source: "n8n" in the body
-		displayName: 'Source',
-		name: 'source',
-		type: 'hidden',
-		default: 'n8n',
-		displayOptions: {
-			show: showOnlyForCommentCreate,
-		},
-		routing: {
-			send: {
-				type: 'body',
-				property: 'source',
-			},
-		},
 	},
 ];